WithFlinkContainers extracted into flink-test-utils
