+++ engine/flink/test-utils/src/main/scala/pl/touk/nussknacker/engine/flink/test/docker/WithFlinkContainers.scala
@@ -1,0 +1,92 @@
+package pl.touk.nussknacker.engine.flink.test.docker
+
+import com.dimafeng.testcontainers.{GenericContainer, LazyContainer}
+import com.typesafe.scalalogging.LazyLogging
+import org.scalatest.Suite
+import org.testcontainers.containers.BindMode
+import org.testcontainers.containers.wait.strategy.LogMessageWaitStrategy
+import org.testcontainers.images.builder.ImageFromDockerfile
+import pl.touk.nussknacker.engine.util.ResourceLoader
+import pl.touk.nussknacker.engine.util.config.ScalaMajorVersionConfig
+
+import java.nio.file.attribute.{PosixFilePermission, PosixFilePermissions}
+import java.nio.file.{Files, Path}
+import java.util.Arrays.asList
+import scala.jdk.CollectionConverters._
+
+// TODO Current Flink setup requires Kafka for zookpeeper - we should run Flink in some lighter version or extract another container with ZK only
+trait WithFlinkContainers extends WithKafkaContainer { self: Suite with LazyLogging =>
+
+  protected val FlinkJobManagerRestPort = 8081
+
+  protected lazy val taskManagerSlotCount = 8
+
+  protected def jobManagerRestUrl =
+    s"http://${jobManagerContainer.container.getHost}:${jobManagerContainer.container.getMappedPort(FlinkJobManagerRestPort)}"
+
+  protected def flinkContainers: List[LazyContainer[_]] = List(jobManagerContainer, taskManagerContainer)
+
+  protected val jobManagerContainer: GenericContainer = {
+    logger.debug(s"Running with number TASK_MANAGER_NUMBER_OF_TASK_SLOTS=$taskManagerSlotCount")
+    val savepointDir = prepareSavepointVolumeDir()
+    new GenericContainer(
+      dockerImage = prepareFlinkImage(),
+      command = "jobmanager" :: Nil,
+      exposedPorts = FlinkJobManagerRestPort :: Nil,
+      env = Map(
+        "SAVEPOINT_DIR_NAME"                -> savepointDir.getFileName.toString,
+        "FLINK_PROPERTIES"                  -> s"state.savepoints.dir: ${savepointDir.toFile.toURI.toString}",
+        "TASK_MANAGER_NUMBER_OF_TASK_SLOTS" -> taskManagerSlotCount.toString
+      ),
+      waitStrategy = Some(new LogMessageWaitStrategy().withRegEx(".*Recover all persisted job graphs.*"))
+    ).configure { self =>
+      self.withNetwork(network)
+      self.setNetworkAliases(asList("jobmanager"))
+      self.withLogConsumer(logConsumer.withPrefix("jobmanager"))
+      self.withFileSystemBind(savepointDir.toString, savepointDir.toString, BindMode.READ_WRITE)
+    }
+  }
+
+  protected val taskManagerContainer: GenericContainer = {
+    new GenericContainer(
+      dockerImage = prepareFlinkImage(),
+      command = "taskmanager" :: Nil,
+      env = Map("TASK_MANAGER_NUMBER_OF_TASK_SLOTS" -> taskManagerSlotCount.toString),
+      waitStrategy = Some(new LogMessageWaitStrategy().withRegEx(".*Successful registration at resource manager.*"))
+    ).configure { self =>
+      self.setNetwork(network)
+      self.setNetworkAliases(asList("taskmanager"))
+      self.withLogConsumer(logConsumer.withPrefix("taskmanager"))
+    }
+  }
+
+  private def prepareFlinkImage(): ImageFromDockerfile = {
+    List("Dockerfile", "entrypointWithIP.sh", "conf.yml", "log4j-console.properties").foldLeft(
+      new ImageFromDockerfile()
+    ) { case (image, file) =>
+      val resource = ResourceLoader.load(s"/docker/$file")
+
+      val flinkLibTweakCommand = ScalaMajorVersionConfig.scalaMajorVersion match {
+        case "2.12" => ""
+        case "2.13" =>
+          s"""
+             |RUN rm $$FLINK_HOME/lib/flink-scala*.jar
+             |RUN wget https://repo1.maven.org/maven2/pl/touk/flink-scala-2-13_2.13/1.1.1/flink-scala-2-13_2.13-1.1.1-assembly.jar -O $$FLINK_HOME/lib/flink-scala-2-13_2.13-1.1.1-assembly.jar
+             |RUN chown flink $$FLINK_HOME/lib/flink-scala-2-13_2.13-1.1.1-assembly.jar
+             |""".stripMargin
+        case v => throw new IllegalStateException(s"unsupported scala version: $v")
+      }
+      val withFlinkLibTweaks = resource.replace("${scala.version.flink.tweak.commands}", flinkLibTweakCommand)
+
+      image.withFileFromString(file, withFlinkLibTweaks)
+    }
+  }
+
+  private def prepareSavepointVolumeDir(): Path = {
+    Files.createTempDirectory(
+      "nussknackerFlinkSavepointTest",
+      PosixFilePermissions.asFileAttribute(PosixFilePermission.values().toSet[PosixFilePermission].asJava)
+    )
+  }
+
+}
